moved functions to utils, further exceptions refactor
